# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HalkYemek API
  version: dev
  description: Beylikdüzü odaklı QR + Wallet yemek sistemi
paths:
...
  /api/v1/businesses/{business_id}/operations/consume-history/:
    get:
      operationId: business_consume_history
      parameters:
      - in: path
        name: business_id
        schema:
          type: integer
        required: true
      - in: query
        name: checkout_status
        schema:
          enum:
          - PENDING
          - CONFIRMED
          - CONSUMED
          - EXPIRED
          - CANCELLED
          type: string
          minLength: 1
        description: |-
          * `PENDING` - PENDING
          * `CONFIRMED` - CONFIRMED
          * `CONSUMED` - CONSUMED
          * `EXPIRED` - EXPIRED
          * `CANCELLED` - CANCELLED
      - in: query
        name: consumed_after
        schema:
          type: string
          format: date-time
      - in: query
        name: consumed_before
        schema:
          type: string
          format: date-time
      tags:
      - business-operations
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '403':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '404':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/businesses/{business_id}/operations/dashboard-summary/:
    get:
      operationId: business_dashboard_summary
      parameters:
      - in: path
        name: business_id
        schema:
          type: integer
        required: true
      tags:
      - business-operations
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '403':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '404':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/businesses/{business_id}/operations/orders/{order_id}/:
    get:
      operationId: business_order_detail
      parameters:
      - in: path
        name: business_id
        schema:
          type: integer
        required: true
      - in: path
        name: order_id
        schema:
          type: integer
        required: true
      tags:
      - business-operations
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '403':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '404':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/businesses/{business_id}/operations/profile/:
    get:
      operationId: business_profile_operations_detail
      parameters:
      - in: path
        name: business_id
        schema:
          type: integer
        required: true
      tags:
      - business-operations
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '403':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '404':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
    patch:
      operationId: business_profile_operations_update
      parameters:
      - in: path
        name: business_id
        schema:
          type: integer
        required: true
      tags:
      - business-operations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedBusinessProfileOperationsUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedBusinessProfileOperationsUpdateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedBusinessProfileOperationsUpdateRequest'
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '403':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '404':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
components:
  schemas:
...
    ListingTypeEnum:
      enum:
      - CONTRACTED
      - VOLUNTEER
      type: string
      description: |-
        * `CONTRACTED` - Contracted
        * `VOLUNTEER` - Volunteer
...
    PatchedBusinessProfileOperationsUpdateRequest:
      type: object
      properties:
        short_description:
          type: string
          maxLength: 280
        intro_text:
          type: string
        badge_text:
          type: string
          maxLength: 64
        marketplace_is_visible:
          type: boolean
        listing_type:
          $ref: '#/components/schemas/ListingTypeEnum'
        is_featured:
          type: boolean
        display_priority:
          type: integer
          minimum: 0
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...